requirements.txt
# Framework y servidor web
fastapi==0.68.0
uvicorn==0.15.0

# gRPC y herramientas (versiones binarias)
grpcio-cp313-win_amd64==1.54.2
protobuf==3.20.3

# CLI y utilidades
typer==0.4.0
requests==2.26.0

# Versiones precompiladas para Python 3.13
aiohttp-cp313-win_amd64==3.8.1
psutil-cp313-win_amd64==5.8.0

# Dependencias adicionales
pydantic==1.10.12
starlette==0.14.2
yarl==1.9.2
multidict==6.0.4
async-timeout==4.0.3
attrs==23.1.0
chardet==4.0.0
